+++ com.unity.shadergraph/Editor/Drawing/Views/StickyNote.cs
@@ -2,8 +2,6 @@ using System.Collections.Generic;
 using UnityEditor.Experimental.GraphView;
 using UnityEngine;
 using UnityEngine.UIElements;
-using System.Reflection;
-using System.Linq;
 
 namespace UnityEditor.ShaderGraph.Drawing
 {
@@ -466,9 +464,18 @@ namespace UnityEditor.ShaderGraph.Drawing
 
                 evt.menu.AppendAction("Fit To Text", OnFitToText, e => DropdownMenuAction.Status.Normal);
                 evt.menu.AppendSeparator();
+                
+                evt.menu.AppendAction("Delete",  OnDelete, e => DropdownMenuAction.Status.Normal);
+                evt.menu.AppendSeparator();
             }
         }
 
+        void OnDelete(DropdownMenuAction menuAction)
+        {
+            m_Graph.owner.RegisterCompleteObjectUndo("Delete Sticky Note");
+            m_Graph.RemoveStickyNote(userData);
+        }
+        
         void OnTitleChange(EventBase e)
         {
             //m_Graph.owner.RegisterCompleteObjectUndo("Title Changed");