+++ com.unity.shadergraph/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [7.2.0] - 2019-XX-XX
+### Added
+- You can now use the right-click context menu to delete Sticky Notes.
 
 ## [7.1.1] - 2019-09-05
 ### Added